Refactor multiple files
