+++ backend/app/routes/broadcast/@validationSchema/index.js
@@ -1,14 +1,16 @@
 const Joi = require('joi');
 
-const addBroadcastValidation = Joi.object().keys({
+const postBroadcastValidationSchema = Joi.object().keys({
   title: Joi.string().required(),
   content: Joi.string().required(),
   link: Joi.string().uri().required(),
-  isExpired: Joi.boolean().required(),
+  expiresOn: Joi.date()
+    .min(new Date(new Date() - 100000))
+    .required(),
   imageUrl: Joi.array().min(1).items(Joi.string().uri()).required(),
-  category: Joi.array().min(1).items(Joi.string()).required(),
+  tags: Joi.array().min(1).items(Joi.string()).required(),
 });
 
 module.exports = {
-  addBroadcastValidation,
+  postBroadcastValidationSchema,
 };
