+++ backend/app/routes/broadcast/index.js
@@ -2,9 +2,9 @@ const router = require('express').Router({ mergeParams: true });
 const validationMiddleware = require('../../../helpers/middlewares/validation');
 const { authMiddleware } = require('../../../helpers/middlewares/auth');
 
-const { addBroadcastValidation } = require('./@validationSchema');
-const addBroadcast = require('./addBroadcast');
+const { postBroadcastValidationSchema } = require('./@validationSchema');
+const postBroadcast = require('./postBroadcast');
 
-router.post('/', validationMiddleware(addBroadcastValidation), authMiddleware, addBroadcast);
+router.post('/', validationMiddleware(postBroadcastValidationSchema), authMiddleware, postBroadcast);
 
 module.exports = router;
